- tạo luồng chạy cơ bản trong g_i và để nội dung cũ vào file mới snake
- tạo file coordinate

- thay đổi file adie cho phù hợp với coordinate và thêm khối if __name__ == '__main__' để import được từ luồng chính g_i

- thêm file dtcd và import nó từ luồng chính g_i
